change doctest to be more readible, move helpers back from tempdir
