+++ tempdir/src/lib.rs
@@ -35,23 +35,3 @@ impl Drop for TempDir {
         self.clear();
     }
 }
-
-// Some helpers
-pub fn get_dir_files_hashset(dir: &str) -> HashSet<String> {
-    let mut files = HashSet::new();
-    for file in read_dir(dir).unwrap() {
-        let filename = file.unwrap().file_name().to_str().unwrap().to_string();
-        files.insert(filename);
-    }
-    files
-}
-
-pub fn assert_correct_files(dir: &str, log_filenames: Vec<&str>) {
-    // TODO: change to ref of vec, prob doesn't need ownership
-    // TODO: fix this complete shitshow
-    let log_files: HashSet<String> = get_dir_files_hashset(dir);
-    let log_files_str: HashSet<&str> = log_files.iter().map(AsRef::as_ref).collect();
-    let expected: HashSet<&str> = log_filenames.into_iter().collect();
-
-    assert_eq!(log_files_str, expected);
-}
